Work on sampling point to support the old and new backend of dpf composites
